+++ ricaun.Revit.DB.Shape.Tests/Color_Tests.cs
@@ -1,6 +1,7 @@
 ﻿using Autodesk.Revit.DB;
 using NUnit.Framework;
 using ricaun.Revit.DB.Shape.Tests.Utils;
+using System.Collections.Generic;
 
 namespace ricaun.Revit.DB.Shape.Tests
 {
@@ -14,22 +15,44 @@ namespace ricaun.Revit.DB.Shape.Tests
         Color blue = new Color(0, 0, 255);
         Color magenta = new Color(255, 0, 255);
         Color white = new Color(255, 255, 255);
-        Color darkgray = new Color(65, 65, 65);
+        Color blackgray = new Color(65, 65, 65);
         Color gray = new Color(128, 128, 128);
+
+        private static Dictionary<string, System.Windows.Media.Color> GetMediaColors()
+        {
+            var result = new Dictionary<string, System.Windows.Media.Color>();
+            foreach (var property in typeof(System.Windows.Media.Colors).GetProperties())
+            {
+                if (property.Name == "Transparent") continue;
+                if (property.GetValue(null) is System.Windows.Media.Color color)
+                    result.Add(property.Name, color);
+            }
+            return result;
+        }
 
         [Test]
         public void Colors_Tests()
         {
-            Assert.IsTrue(Colors.Black.ColorEquals(black));
-            Assert.IsTrue(Colors.Red.ColorEquals(red));
-            Assert.IsTrue(Colors.Yellow.ColorEquals(yellow));
-            Assert.IsTrue(Colors.Green.ColorEquals(green));
-            Assert.IsTrue(Colors.Cyan.ColorEquals(cyan));
-            Assert.IsTrue(Colors.Blue.ColorEquals(blue));
-            Assert.IsTrue(Colors.Magenta.ColorEquals(magenta));
-            Assert.IsTrue(Colors.White.ColorEquals(white));
-            Assert.IsTrue(Colors.DarkGray.ColorEquals(darkgray));
-            Assert.IsTrue(Colors.Gray.ColorEquals(gray));
+            var colorsType = typeof(Colors);
+            var count = colorsType.GetProperties().Length;
+            var mediaColors = GetMediaColors();
+
+            System.Console.WriteLine($"Colors: {count} \tMedia.Colors: {mediaColors.Count}");
+
+            foreach (var color in mediaColors)
+            {
+                var name = color.Key;
+                var expectedColor = new Color(color.Value.R, color.Value.G, color.Value.B);
+                if (colorsType.GetProperty(name)?.GetValue(null) is Color colorTest)
+                {
+                    Assert.IsTrue(colorTest.ColorEquals(expectedColor), 
+                        $"{name} Color ({colorTest.Red},{colorTest.Green},{colorTest.Blue}) is not equal to ({expectedColor.Red},{expectedColor.Green},{expectedColor.Blue})");
+                }
+                else
+                {
+                    Assert.Fail($"Colors.{name} not found.");
+                }
+            }
         }
 
         [Test]
@@ -43,7 +66,7 @@ namespace ricaun.Revit.DB.Shape.Tests
             Assert.IsTrue(Colors.Index.Color_5.ColorEquals(blue));
             Assert.IsTrue(Colors.Index.Color_6.ColorEquals(magenta));
             Assert.IsTrue(Colors.Index.Color_7.ColorEquals(white));
-            Assert.IsTrue(Colors.Index.Color_8.ColorEquals(darkgray));
+            Assert.IsTrue(Colors.Index.Color_8.ColorEquals(blackgray));
             Assert.IsTrue(Colors.Index.Color_9.ColorEquals(gray));
         }
 
@@ -58,7 +81,7 @@ namespace ricaun.Revit.DB.Shape.Tests
             Assert.IsTrue(Colors.Index.Get(5).ColorEquals(blue));
             Assert.IsTrue(Colors.Index.Get(6).ColorEquals(magenta));
             Assert.IsTrue(Colors.Index.Get(7).ColorEquals(white));
-            Assert.IsTrue(Colors.Index.Get(8).ColorEquals(darkgray));
+            Assert.IsTrue(Colors.Index.Get(8).ColorEquals(blackgray));
             Assert.IsTrue(Colors.Index.Get(9).ColorEquals(gray));
         }